refactor: move auth DTOs to common location
