--- src/auth/dto/login.dto.ts
@@ -1,4 +1,0 @@
-import { loginSchema, LoginInput } from '../../models/user.model';
-
-export const LoginDto = loginSchema;
-export type LoginDto = LoginInput;
