--- src/auth/dto/signup.dto.ts
@@ -1,4 +1,0 @@
-import { signupSchema, SignupInput } from '../../models/user.model';
-
-export const SignupDto = signupSchema;
-export type SignupDto = SignupInput;
