+++ src/models/user.model.ts
@@ -5,17 +5,24 @@ import { User as UserEntity } from '../entities/user.entity';
 export type User = UserEntity;
 export type NewUser = Omit<User, 'createdAt'>;
 
-// Zod schemas for validation
-export const insertUserSchema = z.object({
-  id: z.string().uuid(),
+// Base schemas for reusability
+const baseUserFields = {
   email: z.string().email('Invalid email format'),
   password: z.string().min(6, 'Password must be at least 6 characters long'),
+};
+
+const baseUserFieldsWithId = {
+  id: z.string().uuid(),
+  ...baseUserFields,
+};
+
+// Zod schemas for validation
+export const insertUserSchema = z.object({
+  ...baseUserFieldsWithId,
 });
 
 export const selectUserSchema = z.object({
-  id: z.string().uuid(),
-  email: z.string().email(),
-  password: z.string(),
+  ...baseUserFieldsWithId,
   createdAt: z.date(),
 });
 
@@ -34,17 +41,3 @@ export const updateUserSchema = z
 // Utility types
 export type UserWithoutPassword = Omit<User, 'password'>;
 export type UserResponse = UserWithoutPassword;
-
-// Refined schemas for API validation
-export const loginSchema = z.object({
-  email: z.string().email('Invalid email format'),
-  password: z.string().min(1, 'Password is required'),
-});
-
-export const signupSchema = z.object({
-  email: z.string().email('Invalid email format'),
-  password: z.string().min(6, 'Password must be at least 6 characters long'),
-});
-
-export type LoginInput = z.infer<typeof loginSchema>;
-export type SignupInput = z.infer<typeof signupSchema>;
